+++ src/components/captions-control/index.ts
@@ -1,0 +1,1 @@
+export {CaptionsControl} from './captions-control';