+++ src/components/scrollable/scrollable.tsx
@@ -1,6 +1,6 @@
 import {ComponentChildren, h} from 'preact';
 //@ts-ignore
-import {useState, useRef, useMemo, MutableRef} from 'preact/hooks';
+import {useState, useRef, useMemo, MutableRef, useLayoutEffect} from 'preact/hooks';
 import styles from '../../styles/style.scss';
 
 const SCROLL_BAR_TIMEOUT = 250;
@@ -17,6 +17,7 @@ const Scrollable = ({children, isVertical}: {children: ComponentChildren; isVert
   const ref: MutableRef<HTMLDivElement | null> = useRef<HTMLDivElement | null>(null);
   const [scrolling, setScrolling] = useState<boolean>(false);
   const [scrollTimeoutId, setScrollTimeoutId] = useState<number>(-1);
+  const [scrollableHeight, setScrollableHeight] = useState<string>('');
 
   const handleScroll = (): void => {
     clearTimeout(scrollTimeoutId);
@@ -36,11 +37,25 @@ const Scrollable = ({children, isVertical}: {children: ComponentChildren; isVert
     }
   };
 
+  useLayoutEffect(() => {
+    if (isVertical && ref?.current) {
+      const scrollableEl = ref.current;
+      const parentHeight = scrollableEl.parentElement?.clientHeight;
+      if (parentHeight) {
+        const cs = getComputedStyle(scrollableEl.parentElement);
+        const verticalPadding = parseFloat(cs.paddingTop) + parseFloat(cs.paddingBottom);
+        setScrollableHeight(`${parentHeight - verticalPadding}px`);
+      }
+    }
+  }, []);
+
   const scrollableParams = useMemo(() => (isVertical ? {onScroll: handleScroll} : {onWheel: handleWheel, ref}), [isVertical]);
 
   return (
     <div
       className={`${styles.scrollable} ${scrolling ? styles.scrolling : ''} ${isVertical ? styles.vertical : styles.horizontal}`}
+      style={`${isVertical && scrollableHeight ? `height: ${scrollableHeight}` : ''}`}
+      ref={ref}
       {...scrollableParams}
     >
       {children}