+++ src/components/tooltip/tooltip.tsx
@@ -16,6 +16,8 @@ interface TooltipOwnProps {
   type?: ToolTipPosition;
   maxWidth?: string;
   label: string;
+  strictPosition?: boolean;
+  className?: string;
 }
 
 type TooltipProps = ReduxStateProps & TooltipOwnProps;
@@ -69,7 +71,8 @@ class Tooltip extends Component<TooltipProps & WithEventManagerProps, any> {
    */
   static defaultProps = {
     type: ToolTipType.Top,
-    maxWidth: '240px'
+    maxWidth: '240px',
+    strictPosition: false
   };
 
   /**
@@ -174,6 +177,9 @@ class Tooltip extends Component<TooltipProps & WithEventManagerProps, any> {
    * @returns {string} is in boundaries
    */
   isToolTipInBoundaries(): boolean {
+    if (this.props.strictPosition) {
+      return true;
+    }
     const tooltipBoundingRect = this.textElement.getBoundingClientRect();
     const playerContainerRect = this.props.playerClientRect;
 
@@ -262,6 +268,9 @@ class Tooltip extends Component<TooltipProps & WithEventManagerProps, any> {
    */
   render(props: any): VNode<any> {
     const className = [style.tooltipLabel, style[`tooltip-${this.state.type}`]];
+    if (props.className) {
+      className.push(props.className);
+    }
     this.state.showTooltip && this.state.valid ? className.push(style.show) : className.push(style.hide);
     if (props.isMobile) {
       return toChildArray(props.children)[0] as VNode<any>;
@@ -275,8 +284,7 @@ class Tooltip extends Component<TooltipProps & WithEventManagerProps, any> {
         className={style.tooltip}
         onMouseOver={this.onMouseOver}
         onMouseLeave={this.onMouseLeave}
-        ref={el => (el ? (this.tooltipElement = el) : undefined)}
-      >
+        ref={el => (el ? (this.tooltipElement = el) : undefined)}>
         {children}
         <span style={{maxWidth: props.maxWidth}} ref={el => (el ? (this.textElement = el) : undefined)} className={className.join(' ')}>
           {props.label}