+++ src/utils/index.ts
@@ -1,6 +1,6 @@
 export {bindActions} from './bind-actions';
 export {focusElement} from './focus-element';
-export {KeyMap} from './key-map';
+export {KeyMap, KeyCode} from './key-map';
 export {default as getLogger} from './logger';
 export {withKeyboardA11y} from './popup-keyboard-accessibility';
 export * from './time-format';